Add a TCP client and a TCP server.
Add a TCP client connecting to an arbirarly chosed
ipv4 address and port cobination.
It settles it's connection upon a file descriptor.

Add a TCP server which settles it's connection
as well on a file descriptor. After starting listening
on the socket file descriptor, a select with a timeout
is set to limit the time for which the program will wait for an incomming
connection, thus preventing it from eventual stuck in infinite loop.