Add brute force example of on-file sync
Provide two cases with O_SYNC and O_APPEND.  Each thread (within a
single process) attempts to open an existing file and then writes to
it. Just that.

O_SYNC makes "a write operation to the file to complete according to
the requirements of sync. I/O file integrity completion. By the time
write(2) (or similar), the output data and associated file metadata
have been transferred to the underlying hardware (i.e., as though each
write(2) was followed by a call to fsync(2))." (source: man open) As
the result, there is no synchronisation between the threads and the
text is corrupted in the case of a single char. Although it seems that
only a single thread (presumably the "faster" one?) is able to write a
part of it's text if it's a bigger blob.

O_APPEND "the file is opened in append mode. Before each write(2), the
file offset is positioned at the end o f the file, as if with
lseek(2). The modification of the file offset and the write operation
are performed as a single atomic step."  As the result, there is no
synchronisation between the threads and the text is corrupted in the
case of a single char. Although it seems that only a single thread
(presumably the "faster" one?) is able to write a part of it's text if
it's a bigger blob.