Add a program to list all process' fds
List process' fd in form of: symlink path -> path of linked file To do
that, use opendir(2) to go through the directory, retrieve entries
with readdir(2) and Finlay use readlink(2) to get the path of the
linked file.

To assure that the readlink(2) is reading a symbolic link, check on
the dirent d_type (DT_LNK value).